requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate
Pillow
faiss-cpu
streamlit
openai             # only if you use GPT-4o-mini (cloud)
numpy
tqdm
opencv-python
